requirements.txt
websocket-client>=1.0.0
websocket-server>=0.6.1
